a return to `eager` updates
a (many) few commits ago a change was made to keep a q of literals and to only update the valuation when taking a literal from the q

this is nice and uniform, however it means that existing information when updating the watch for a literal
for, if a literal is q'd and a watch is put to conflict with that literal, then it would have been better to set a different watch, for example

this commit reverses that change

in addition, varaible id's are change to u32, which to some surprise is a measurable difference
and, a few other things are tidied

+++ src/structures/variable.rs
@@ -1,6 +1,6 @@
 use crate::structures::{level::LevelIndex, solve::ClauseKey};
 
-pub type VariableId = usize;
+pub type VariableId = u32;
 use std::cell::UnsafeCell;
 
 pub struct Variable {
